HV-755 Extracting overrides functionality into ReflectionHelper and making use of it in ValidationExtension to properly detect all override cases, especially the ones using generics.
